clean: move losses into separate file
